+++ addon/mixins/flexberry-lookup-controller.js
@@ -76,8 +76,6 @@ export default Ember.Mixin.create(ReloadListMixin, {
   */
   lookupModalWindowPerPage: 5,
 
-  modalIsShow: false,
-
   actions: {
     /**
       Handles action from lookup choose action.