+++ CHANGELOG.md
@@ -5,7 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 * `flexberry-lookup` component: 
-    * Add `lookup-event` servise for to set when modal dialog is open. Now the user can not press the choose button several times before the modal dialog is open.
+    * Add `lookup-events` service for to set when modal dialog is open. Now the user can not press the choose button several times before the modal dialog is open.
+    * Add in `lookup-events` service new triggers: lookupDialogOnShowTrigger, lookupDialogOnVisibleTrigger, lookupDialogOnHiddenTrigger. 
 
 ## [0.8.0-beta.3] - 2016-10-31
 ### Changed